refactor: 루틴 enum 패키지 구조 변경
