+++ src/main/java/com/lived/domain/routine/dto/RoutineDeleteRequestDTO.java
@@ -1,6 +1,6 @@
 package com.lived.domain.routine.dto;
 
-import com.lived.domain.routine.entity.enums.DeleteType;
+import com.lived.domain.routine.enums.DeleteType;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.time.LocalDate;
