+++ src/main/java/com/lived/domain/routine/dto/RoutineTreeResponseDTO.java
@@ -1,6 +1,6 @@
 package com.lived.domain.routine.dto;
 
-import com.lived.domain.routine.entity.enums.FruitType;
+import com.lived.domain.routine.enums.FruitType;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.util.List;
