+++ src/main/java/com/lived/domain/routine/dto/RoutineUpdateRequestDTO.java
@@ -1,7 +1,7 @@
 package com.lived.domain.routine.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.lived.domain.routine.entity.enums.RepeatType;
+import com.lived.domain.routine.enums.RepeatType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
