--- src/main/java/com/lived/domain/routine/entity/RoutineGlobalStats.java
@@ -1,4 +1,0 @@
-package com.lived.domain.routine.entity;
-
-public class RoutineGlobalStats {
-}
