+++ src/main/java/com/lived/domain/routine/service/RoutineStatisticsService.java
@@ -6,9 +6,9 @@ import com.lived.domain.routine.dto.*;
 import com.lived.domain.routine.entity.RoutineBigFruit;
 import com.lived.domain.routine.entity.RoutineFruit;
 import com.lived.domain.routine.entity.RoutineHistory;
-import com.lived.domain.routine.entity.enums.BigFruitType;
-import com.lived.domain.routine.entity.enums.DayStatus;
-import com.lived.domain.routine.entity.enums.FruitType;
+import com.lived.domain.routine.enums.BigFruitType;
+import com.lived.domain.routine.enums.DayStatus;
+import com.lived.domain.routine.enums.FruitType;
 import com.lived.domain.routine.entity.mapping.MemberRoutine;
 import com.lived.domain.routine.enums.StatisticsType;
 import com.lived.domain.routine.repository.MemberRoutineRepository;
